Refactor chip testing utilities and consolidate factory functions
- Updated imports in builder_patterns.rs and chip_integration_tests.rs to use the new unified chip_factory structure.
- Deprecated chip_test_helpers.rs and migrated its functionality to chip_factory with improved organization.
- Enhanced chip_factory.rs by categorizing functions into basic, states, variants, collections, assertions, and generators for better maintainability.
- Updated chip_tests.rs and chip_validation_tests.rs to utilize the new factory functions and assertions.
- Refactored collection_factory.rs to align with the new chip_factory structure.
- Modified mod.rs to re-export specific items from chip_factory and collection_factory to avoid naming conflicts and improve clarity.
